chore(webpage): Move load of plugins to widgets
